+++ R/select_rows.R
@@ -14,8 +14,7 @@ validate_phenotype_definitions = function(pheno, available, csd=NULL) {
       || stringr::str_detect(pheno, 'Total|All'))
     return('')
 
-  phenos = stringr::str_split(pheno, '[,/]')[[1]] %>%
-    stringr::str_trim()
+  phenos = split_quoted(pheno, c(',', '/'))
 
   if (!all(stringr::str_detect(phenos, '^[~#]|[+-]$')))
     return('Phenotype definitions must start with ~ or #, or end with + or -.')
@@ -151,7 +150,7 @@ parse_phenotypes = function(...) {
         stop(paste("Phenotype selectors may not contain both '/' and ',':",
                    pheno))
       ## Split the phenotypes and convert formulae
-      purrr::map(split_and_trim(pheno, '/'), ~{
+      purrr::map(split_quoted(pheno), ~{
         if (startsWith(.x, '~')) stats::as.formula(.x, globalenv()) else .x
       })
     }
@@ -221,6 +220,37 @@ phenotype_columns = function(phenos) {
 split_and_trim = function(str, pattern) {
   stopifnot(is.character(str), length(str)==1)
   stringr::str_trim(stringr::str_split(str, pattern)[[1]])
+}
+
+#' Split a single string on / and trim white space from the results.
+#' Slashes quoted with backquotes are not split.
+#' `str` must not contain tabs
+#' @param str A single string.
+#' @param split Character vector containing characters to split on
+#' @return A character vector of split components.
+#' @keywords internal
+split_quoted = function(str, split='/') {
+  stopifnot(is.character(str), length(str)==1)
+  stopifnot(is.character(split), all(purrr::map_int(split, nchar) ==1))
+
+  # Split into individual characters
+  chars = strsplit(str, split='')[[1]]
+  if ('\t' %in% chars)
+    stop('Tabs not allowed in split_quoted().')
+
+  # Find character locations preceded by an even number of `
+  even_backticks = cumsum(chars=='`') %% 2 == 0
+
+  # Find split characters that are preceded by an even number of `
+  split_locations = (chars %in% split) & even_backticks
+
+  # Use tab as a marker
+  chars[which(split_locations)] = '\t'
+  str2 = paste0(chars, collapse='')
+
+  # Now the / we want to split on have been changed to ,
+  result = stringr::str_split(str2, '\t')[[1]]
+  stringr::str_trim(result)
 }
 
 #' Make user-friendly names for phenotypes